src/requirements.txt
requests
atlassian-python-api
boto3 #no-deploy
botocore #no-deploy